+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 18.2.3
 
+* `Core`:
+  * Add default middleware `ExceptionsThrottlerTelegramBotMiddleware`
+  * Make `TelegramBotMiddlewaresPipelinesHandler` to be default `TelegramBotPipelinesHandler`
+
 ## 18.2.2
 
 * `Version`: